refactor: 모달 모듈에 최상위 디렉터리 lib을 생성하고 폴더 분리
