+++ components/src/lib/assets/close-button.png
