+++ components/src/lib/index.ts
@@ -1,0 +1,2 @@
+export { default as Modal } from './modal/Modal';
+export * from './modal/types';
