--- components/src/modal/index.ts
@@ -1,2 +1,0 @@
-export { default as Modal } from './Modal';
-export * from './types';
