+++ components/vite.config.ts
@@ -6,7 +6,7 @@ import dts from 'vite-plugin-dts';
 export default defineConfig({
   build: {
     lib: {
-      entry: path.resolve(__dirname, 'src/modal/index.ts'),
+      entry: path.resolve(__dirname, 'src/lib/index.ts'),
       name: 'index',
       fileName: 'index',
     },
@@ -25,7 +25,7 @@ export default defineConfig({
   plugins: [
     react(),
     dts({
-      include: ['src/modal'],
+      include: ['src/lib'],
       tsconfigPath: './tsconfig.app.json',
     }),
   ],
